requirements.txt
# Azure Functions Python Requirements
azure-functions>=1.18.0
azure-functions-worker>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests
requests>=2.31.0

# Date and time handling
python-dateutil>=2.8.0

# JSON handling (built-in, but explicit for clarity)
# json - built-in module

# Logging (built-in, but explicit for clarity)
# logging - built-in module

# Type hints (built-in in Python 3.5+)
# typing - built-in module

# Optional: For more advanced data analysis
scipy>=1.10.0
scikit-learn>=1.3.0

# Optional: For Azure integrations
azure-storage-blob>=12.19.0
azure-keyvault-secrets>=4.7.0
azure-identity>=1.15.0

# Optional: For database connections
pyodbc>=4.0.39
sqlalchemy>=2.0.0

# Optional: For monitoring and telemetry
opencensus-ext-azure>=1.1.13
applicationinsights>=0.11.10

